+++ examples/simple/packs/greeting/lib/greeting.rb
@@ -4,7 +4,9 @@ require 'faker'
 
 class Greeting
   def self.hello
-    Faker::Name.name
+    name = Faker::Name.name
+    prefix = ENV['GREETING_PREFIX']
+    prefix ? "#{prefix}, #{name}" : name
   end
 
   def self.faker_version